Camera Movement Setup & Basic Object Spawning Click & Drop
Created a class that moves the camera, another that is responsible for placing gameObjects in the city (not final, just an initial method, triggered by an UI button, then by a Input Action to place the game object - to be changed to a drag and drop instead) and a class for the gameobject 2D image to follow the camera when it is being placed. Plus created a generic gameobject to serve as the city (for now). All part of step 3 of the test.