Working Drag & Drop Implemented
Created a script to handle UI button being pressed, which calls the StartGameObjectPlacement, rather than it being called through the button's inspector OnClick. This way it allows for all of the remaining logic to be handled outside of the button. DragNDropPlacer was also changed to be updated with the wanted prefabs relating to a specific button, removing the need for lists (which I was originally considering). I also corrected one error in CameraControlls script.