Added Audio, Changed SoundManager to be more efficient and fixed small error in DragNDropPlacer
The soundManager was creating the audio sources at runtime and causing a big delay because of it. It now has an audio source pool to handle playing audio clips as needed. Also changed spawn position for in DragNDropPlacer to not go into the city, by using the city normal. NOTE: Sounds still have delay, should change it.